+++ app/src/src_nonfreenet/org/breezyweather/sources/atmo/AtmoApi.kt
@@ -19,6 +19,7 @@ package org.breezyweather.sources.atmo
 import io.reactivex.rxjava3.core.Observable
 import org.breezyweather.sources.atmo.json.AtmoPointResult
 import retrofit2.http.GET
+import retrofit2.http.Header
 import retrofit2.http.Query
 
 /**
@@ -28,7 +29,9 @@ interface AtmoApi {
 
     @GET("point?with_list=true")
     fun getPointDetails(
-        @Query("api_token") apiToken: String,
+        @Header("api_token") headerApiToken: String?,
+        @Header("User-Agent") userAgent: String = "okhttp/3.14.9",
+        @Query("api_token") queryApiToken: String?,
         @Query("x") longitude: Double,
         @Query("y") latitude: Double,
         @Query("datetime_echeance") datetimeEcheance: String,